% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{ct1}
\alias{ct1}
\title{\eqn{C_T} values from qPCR (separate tubes)}
\format{
A data.frame with 12 rows and 2 variables:
\describe{
  \item{c_myc}{\eqn{C_T} values of the target gene \strong{c-myc}}
  \item{GAPDH}{\eqn{C_T} values of the control gene \strong{GAPDH}}
}
}
\source{
\url{http://www3.appliedbiosystems.com/cms/groups/mcb_support/documents/generaldocuments/cms_040980.pdf}
}
\usage{
ct1
}
\description{
A dataset containing the \eqn{C_T} values of two genes from a qPCR experiment.
Samples were prepared from human tissues; Brain and kidney (n = 6) each.
Primers for each genes were run in separate reaction tubes.
}
\seealso{
\code{\link{ct2}}

\code{\link{ct3}}
}
\keyword{datasets}
