% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{ct5}
\alias{ct5}
\title{\eqn{C_T} values from qPCR with multiple reference genes}
\format{A data.frame with 4 rows and 4 variables:
\describe{
  \item{GOT1}{\eqn{C_T} values of the gene of interest}
  \item{REF1}{\eqn{C_T} values of the first reference gene}
  \item{REF2}{\eqn{C_T} values of the second reference gene}
  \item{REF3}{\eqn{C_T} values of the third reference gene}
}}
\usage{
ct5
}
\description{
A dataset containing the \eqn{C_T} values of one gene of interest and three
reference genes in two control and two treatment samples.
}
\keyword{datasets}
