% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCR.R
\name{data_generation}
\alias{data_generation}
\title{Positive-definite covariance regression (PCR)}
\usage{
data_generation(n, p, T_0, rho_1 = NULL, rho_2 = NULL, rho_0 = NULL)
}
\arguments{
\item{n}{an integer represents the sample size}

\item{p}{an integer represents the dimension of the response}

\item{T_0}{an iterger represents the replicated observations for the response}

\item{rho_1}{correlation in matrix \eqn{B_1}}

\item{rho_2}{correlation in matrix \eqn{B_2}}

\item{rho_0}{the dependence between different observations of the response}
}
\value{
a list of three variables, the covariates matrix, the response array and the true coefficient matirces \eqn{B_1} and \eqn{B_2}
\item{Z}{a n x p matrix refers to observations of the covariates}
\item{Y}{a T_0 x p x n array refers to the observations of the response}
\item{B_list}{a p x p x 2 array with each element be the coefficient matrix in the regression model}
}
\description{
Positive-definite covariance regression (PCR)
}
