% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{.pcr_amount}
\alias{.pcr_amount}
\title{Calculate the amounts}
\usage{
.pcr_amount(vec, a, b)
}
\arguments{
\item{vec}{A vector of numerics}

\item{a}{A numeric}

\item{b}{A numeric}
}
\value{
A vector of numerics
}
\description{
Calculate the amounts
}
\examples{
vec <- rnorm(6, 30, 1)
pcr:::.pcr_amount(vec, 1, 1)
}
\keyword{internal}
