% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{.pcr_average}
\alias{.pcr_average}
\title{Get vector average by a variable}
\usage{
.pcr_average(vec, var)
}
\arguments{
\item{vec}{A vector of numerics}

\item{var}{A grouping variable}
}
\value{
A vector of numerics
}
\description{
Get vector average by a variable
}
\examples{
vec <- rnorm(6, 30, 1)
var <- rep(c('group1', 'group2'), 3)
pcr:::.pcr_average(vec, var)

}
\keyword{internal}
