% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{.pcr_calibrate}
\alias{.pcr_calibrate}
\title{Calibrate values by a row}
\usage{
.pcr_calibrate(df, reference_group, mode = "subtract", tidy = FALSE)
}
\arguments{
\item{df}{A data.frame}

\item{reference_group}{A character string of the the entery in the rows of a
grouping variable}

\item{mode}{A character string of the normalization mode to be used. Default
is 'subtract'. Other possible modes include 'divide'}

\item{tidy}{A logical, default FALSE. When TRUE returns a tidy data.frame}
}
\value{
A data.frame of the same dimensions after subtracting or dividing by
a reference_group/row. The function returns ignores non numeric columns.
When tidy is TRUE returns a tidy data.frame with the columns: gene and
calibrated as well as any non numeric columns such as a grouping variable
group/amount.
}
\description{
Uses subtraction or division to caliberate values in all rows to a sepcified
row
}
\details{
Used to calibrate average ct or input amounts by a reference_group/row
}
\examples{
# locate and read raw ct data
fl <- system.file('extdata', 'ct1.csv', package = 'pcr')
ct1 <- readr::read_csv(fl)

# add grouping variable
group_var <- rep(c('brain', 'kidney'), each = 6)

# calculate averages
ave <- .pcr_average(ct1, group_var = group_var)

# calculate delta ct
dct <- .pcr_normalize(ave, 'GAPDH')

# calculate delta delta ct
.pcr_calibrate(dct, 'brain', tidy = TRUE)

# calculate delta delta ct and return a tidy data.frame
.pcr_calibrate(dct, 'brain', tidy = TRUE)

}
\keyword{internal}
