% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{.pcr_error}
\alias{.pcr_error}
\title{Propage two vectors}
\usage{
.pcr_error(vec, ref)
}
\arguments{
\item{vec}{A vector of numerics}

\item{ref}{A numeric vector}
}
\value{
A vector of numerics
}
\description{
Propage two vectors
}
\examples{
vec <- rnorm(6, 30, 1)
ref <- rnorm(6, 30, .1)
pcr:::.pcr_error(vec, ref)
}
\keyword{internal}
