% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{.pcr_relative}
\alias{.pcr_relative}
\title{Raise two to a vector power}
\usage{
.pcr_relative(vec)
}
\arguments{
\item{vec}{A vector of numerics}
}
\value{
A numeric
}
\description{
Raise two to a vector power
}
\examples{
vec <- rnorm(6, 30, 1)
pcr:::.pcr_relative(vec)
}
\keyword{internal}
