% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{.pcr_rsquared}
\alias{.pcr_rsquared}
\title{Calculate R squared}
\usage{
.pcr_rsquared(vec, var)
}
\arguments{
\item{vec}{A vector of numerics}

\item{var}{A grouping variable}
}
\value{
A numeric
}
\description{
Calculate R squared
}
\examples{
vec <- rnorm(6, 30, 1)
var <- rep(c(.1, .5), 3)
pcr:::.pcr_rsquared(vec, var)

}
\keyword{internal}
