% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{.pcr_slope}
\alias{.pcr_slope}
\title{Calculate the slope of a line}
\usage{
.pcr_slope(vec, var)
}
\arguments{
\item{vec}{A vector of numerics}

\item{var}{A grouping variable}
}
\value{
A numeric
}
\description{
Calculate the slope of a line
}
\examples{
vec <- rnorm(6, 30, 1)
var <- rep(c(.1, .5), 3)
pcr:::.pcr_slope(vec, var)

}
\keyword{internal}
