% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{.pcr_trend}
\alias{.pcr_trend}
\title{Calculate the linear trend}
\usage{
.pcr_trend(df, amount)
}
\arguments{
\item{df}{A data.frame of raw ct values or the delta ct values calculated
by \link{.pcr_normalize}}

\item{amount}{A numeric vector input amounts/dilutions of legnth equals the
number of the rows of df.}
}
\value{
A data.frame of 4 columns
\itemize{
  \item gene The column names of df
  \item intercept The intercept of the line
  \item slope The slope of teh line
  \item r_squared The squared correlation
}
}
\description{
Calculates the linear trend; intercept and slope between two variables
}
\details{
Used to calculate the linear trend; intercept and slope for a line
between each column of ct or delta ct values and the log10 input amount
}
\examples{
# locate and read file
fl <- system.file('extdata', 'ct3.csv', package = 'pcr')
ct3 <- read.csv(fl)

# make amount/dilution variable
amount <- rep(c(1, .5, .2, .1, .05, .02, .01), each = 3)

# calculate trend
.pcr_trend(ct3, amount = amount)

}
\keyword{internal}
