% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{effeciency}
\alias{effeciency}
\title{PCR effecienty calculated from pcr_dilute_*}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 5 columns.}
\usage{
effeciency
}
\description{
PCR effecienty calculated from pcr_dilute_*
}
\keyword{datasets}
