% Generated by roxygen2: do not edit by hand

\name{eq_clean_date}
\alias{eq_clean_date}
\title{Clean earthquakE dates}
\usage{
eq_clean_date(df)
}
\arguments{
\item{df}{A data.frame such as this obtainded from NOAA.}
}
\value{
A data.frame with a tidy date column.
}
\description{
Clean earthquake dates
}
\examples{
\dontrun{
file_path <- system.file("data", "dt.tsv", package = "PCR")
signif <- read_tsv(file_path) \%>\%
  eq_clean_date \%>\%
}
}
