% Generated by roxygen2: do not edit by hand
\name{eq_clean_location}
\alias{eq_clean_location}
\title{Clean earthquak locations}
\usage{
eq_clean_location(df)
}
\arguments{
\item{df}{A data.frame such as this obtainded from NOAA.}
}
\value{
A data.frame with a tidy location column
}
\description{
Clean earthquak locations
}
\examples{
\dontrun{
file_path <- system.file("data", "dt.tsv", package = "PCR")
signif <- read_tsv(file_path) \%>\%
  eq_clean_date \%>\%
  eq_clean_location
}
}
