% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create earthquake label}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{A data.frame such as this obtainded from NOAA.}
}
\value{
A column of earthquake annotation; locaion, magnitude and deaths. Ignores the fiels when NA.
}
\description{
Create earthquake label
}
\examples{
\dontrun{
# load and clean data
file_path <- system.file("data", "dt.tsv", package = "PCR")
dt <- read_tsv(file_path) \%>\%
  eq_clean_date \%>\%
  eq_clean_location

# make graph  
dt \%>\%
filter(COUNTRY == 'MEXICO' & year(date) >= 2000) \%>\%
  mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map()
}
}
