% Generated by roxygen2: do not edit by hand
\name{eq_map}
\alias{eq_map}
\title{Map earthquake data to a leaflet tile}
\usage{
eq_map(df)
}
\arguments{
\item{df}{A data.frame such as this obtainded from NOAA.}
}
\value{
A graph with markers for specified earthquaks and popups with info; location, magnitude and deaths.
}
\description{
Map earthquake data to a leaflet tile
}
\examples{
\dontrun{
# load and clean data
file_path <- system.file("data", "dt.tsv", package = "PCR")
dt <- read_tsv(file_path) \%>\%
  eq_clean_date \%>\%
  eq_clean_location

# make graph
dt \%>\%
filter(COUNTRY == 'MEXICO' & year(date) >= 2000) \%>\%
  mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map()
}
}
