% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{hk_tissue}
\alias{hk_tissue}
\title{Expression values reference genes}
\format{A \code{data.frame} with 85 rows and 11 variables}
\usage{
hk_tissue
}
\description{
A dataset containing the relative expression values of ten reference genes
in five different tissues.
}
\keyword{datasets}
