% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{pcr1_ct}
\alias{pcr1_ct}
\title{PCR separate wells experiment raw data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 2 columns.}
\usage{
pcr1_ct
}
\description{
PCR separate wells experiment raw data
}
\keyword{datasets}
