% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{pcr1_norm}
\alias{pcr1_norm}
\title{PCR separate wells experiment processed data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 9 columns.}
\usage{
pcr1_norm
}
\description{
PCR separate wells experiment processed data
}
\keyword{datasets}
