% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{pcr_amounts}
\alias{pcr_amounts}
\title{PCR amount using standard curve}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 7 columns.}
\usage{
pcr_amounts
}
\description{
PCR amount using standard curve
}
\keyword{datasets}
