% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_functions.R
\name{pcr_analyze}
\alias{pcr_analyze}
\title{pcr_analyze}
\usage{
pcr_analyze(df, group_var, reference_gene, reference_group, intervals = TRUE,
  method = "delta_delta_ct", mode = "average_ct")
}
\arguments{
\item{df}{A data.frame of numeric ct value of n genes in columns and m
samples in raws.}

\item{group_var}{A vector of length m as a grouping variables of samples}

\item{reference_gene}{A character string of the name of the column
correspoinding to the reference gene}

\item{reference_group}{A character string of the reference group as it is
recorded in the grouping variable}

\item{intervals}{A logical (default TRUE) to whether or not to calculate the
error intervals}

\item{method}{A character string. Default is "delta_delta_ct"}

\item{mode}{A character string. Possible inputs are "average_ct" or
"average_dct"}
}
\value{
A tidy data.frame of calculated expression values and errors
}
\description{
A wrapper function to perform different analysis methods and modes
}
\examples{
# locate and read raw ct data
fl <- system.file('extdata', 'pcr1_ct.csv', package = 'pcr')
pcr1_ct <- readr::read_csv(fl)

# add grouping variable
group_var <- rep(c('brain', 'kidney'), each = 6)

# calculate all values and errors in one step
pcr_analyze(pcr1_ct,
group_var = rep(c('brain', 'kidney'), each = 6),
reference_gene = 'GAPDH',
reference_group = 'brain')

}
