% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_functions.R
\name{pcr_assess}
\alias{pcr_assess}
\title{Assess PCR data quality}
\usage{
pcr_assess(df, amount, mode = "effeciency", plot = FALSE)
}
\arguments{
\item{df}{A data.frame of exactly two columns containing ct values of a
reference gene and another from an experiment run with differen dilutions
(RNA amounts)}

\item{amount}{A numeric vector of length equals nrow df with RNA amounts}

\item{mode}{A character string of assessment mode. Default "effeciency"}

\item{plot}{A logical default FALSE of whether plot or return the data}
}
\value{
A data.frame of 5 columns
}
\description{
Assess PCR data quality
}
\examples{
# locate and read data
fl <- system.file('extdata', 'pcr_dilute.csv', package = 'pcr')
pcr_dilute <- readr::read_csv(fl)

# make a vector of RNA amounts
amount <- rep(c(1, .5, .2, .1, .05, .02, .01), each = 3)

# calculate effeciencey
pcr_assess(pcr_dilute,
           amount = amount)

}
