% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{pcr_average}
\alias{pcr_average}
\title{Calculate average ct value}
\usage{
pcr_average(df, group_var, amount, tidy = FALSE)
}
\arguments{
\item{df}{A data.frame of numeric ct value of n genes in columns and m
samples in raws}

\item{group_var}{A vector of length m as a grouping variables of samples}

\item{amount}{A numeric vector of the input dilutions or amounts}

\item{tidy}{A logical, default FALSE. When TRUE returns a tidy data.frame}
}
\value{
A data.frame with a column for each and an addition colum for the,
grouping variable that was provided, and a raw for each unique the  entry
in the group_var or amount. When tidy is TRUE the function returns a tidy
data.frame with three columns; group/amount, gene and average.
}
\description{
Takes a data.frame of raw ct values and returns the averages in different
groups or dilutions
}
\examples{
# using a group_var variabale
# locate and read raw ct data
fl <- system.file('extdata', 'ct1.csv', package = 'pcr')
ct1 <- readr::read_csv(fl)

# add grouping variable
group_var <- rep(c('brain', 'kidney'), each = 6)

# calculate averages
pcr_average(ct1, group_var = group_var)

# calculate averages and return a tidy data.frame
pcr_average(ct1, group_var = group_var, tidy = TRUE)

# using a amount variable
# locate and read raw ct data
fl <- system.file('extdata', 'ct3.csv', package = 'pcr')
ct3 <- readr::read_csv(fl)

# add amount variable
amount <- amount <- rep(c(1, .5, .2, .1, .05, .02, .01), each = 3)

# calculate averages
pcr_average(ct3, amount = amount)

# calculate averages and return a tidy data.frame
pcr_average(ct3, amount = amount, tidy = TRUE)

}
