% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_functions.R
\name{pcr_caliberate}
\alias{pcr_caliberate}
\title{Caliberate genes to a reference group}
\usage{
pcr_caliberate(df, group_var, reference_group, intervals = TRUE,
  method = "delta_ct", mode = "average_ct")
}
\arguments{
\item{df}{A data.frame of numeric ct value of n genes in columns and m
samples in raws.}

\item{group_var}{A vector of length m as a grouping variables of samples}

\item{reference_group}{A character string of the reference group as it is
recorded in the grouping variable}

\item{intervals}{A logical (default TRUE) to whether or not to calculate the
error intervals}

\item{method}{A character string of calculation method. Default "delta_ct"}

\item{mode}{A character string. Possible inputs are "average_ct" or
"average_dct"}
}
\value{
A data.frame of caliberated, relative values and errors
}
\description{
Caliberate ct values of genes to a reference group and obtain the fold-
change usually for reference genes
}
\examples{
# locate and read file
fl <- system.file('extdata', 'pcr1_ct.csv', package = 'pcr')
pcr1_ct <- readr::read_csv(fl)

# make a data.frame of two identical columns
pcr_hk <- data.frame(
  GAPDH1 = pcr1_ct$GAPDH,
  GAPDH2 = pcr1_ct$GAPDH
  )

# add grouping variable
group_var <- rep(c('brain', 'kidney'), each = 6)

# calculate caliberation
pcr_caliberate(pcr_hk,
               group_var = group_var,
               reference_group = 'brain')

}
