% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses_fun.R
\name{pcr_curve}
\alias{pcr_curve}
\title{Calculate the standard_curve}
\usage{
pcr_curve(df, group_var, reference_gene, reference_group,
  mode = "average_amounts", intercept, slope)
}
\arguments{
\item{df}{A data.frame of ct values}

\item{group_var}{A character vector of a grouping variable}

\item{reference_gene}{A character string of the column name of a control gene}

\item{reference_group}{A character string of the control group}

\item{mode}{A character string; 'average_amounts' or 'average_normalized'}

\item{intercept}{A numerice vector of intercepst and length equals the number of genes}

\item{slope}{A numerice vector of slopes length equals the number of genes}
}
\value{
A data.frame of 7 columns
\itemize{
  \item group
  \item gene
  \item normalized
  \item caliberated
  \item error
  \item lower
  \item upper
}
}
\description{
Calculate the standard_curve
}
\examples{
# locate and read file
fl <- system.file('extdata', 'ct3.csv', package = 'pcr')
ct3 <- readr::read_csv(fl)

fl <- system.file('extdata', 'ct1.csv', package = 'pcr')
ct1 <- readr::read_csv(fl)

# make a vector of RNA amounts
amount <- rep(c(1, .5, .2, .1, .05, .02, .01), each = 3)

# calculate curve
standard_curve <- pcr_assess(ct3, amount = amount, method = 'standard_curve')
intercept <- standard_curve$intercept
slope <- standard_curve$slope

# make grouping variable
group <- rep(c('brain', 'kidney'), each = 6)

# apply the standard curve method
pcr_ddct(ct1,
         group_var = group,
         reference_gene = 'GAPDH',
         reference_group = 'brain')

}
