% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_functions.R
\name{pcr_cv}
\alias{pcr_cv}
\title{Calculates cv}
\usage{
pcr_cv(df, group_var, reference_gene)
}
\arguments{
\item{df}{A data.frame of numeric ct value of n genes in columns and m
samples in raws.}

\item{group_var}{A vector of length m as a grouping variables of samples}

\item{reference_gene}{A character string of the name of the column
correspoinding to the reference gene}
}
\value{
A data.frame of the cv term for each gene in each sample
}
\description{
Error terms for the standard curve relative quantification
}
\examples{
# locate and read data
fl <- system.file('extdata', 'pcr_dilute.csv', package = 'pcr')
pcr_dilute <- readr::read_csv(fl)

fl <- system.file('extdata', 'pcr1_ct.csv', package = 'pcr')
pcr1_ct <- readr::read_csv(fl)

# make a vector of RNA amounts
amount <- rep(c(1, .5, .2, .1, .05, .02, .01), each = 3)

# calculate curve
standard_curve <- pcr_assess(pcr_dilute, amount = amount, mode = 'standard_curve')
intercept <- standard_curve$intercept
slope <- standard_curve$slope

# calculate amounts
input_amounts <- pcr_amount(pcr1_ct,
                            intercept = intercept,
                            slope = slope)

# make grouping variable
group <- rep(c('brain', 'kidney'), each = 6)

# calculate cv errors
pcr_cv(input_amounts,
       group_var = group,
       reference_gene = 'GAPDH')

}
