% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_functions.R
\name{pcr_dct}
\alias{pcr_dct}
\title{Calculate delta ct value}
\usage{
pcr_dct(df, reference_gene)
}
\arguments{
\item{df}{A data.frame of average ct values for each gene in each conditon
in addition to a grouping variable such as the output of \link{pcr_ave}}

\item{reference_gene}{A character string of the name of the column
correspoinding to the reference gene}
}
\value{
A data.frame of delta ct values for each gene and a grouping
variable. The column corresponding to the reference gene is dropped.
}
\description{
Takes a data.frame of average ct values for each gene in each condition and
returns the delta ct value (average gene of interest - average reference
gene)
}
\examples{
# locate and read raw ct data
fl <- system.file('extdata', 'pcr1_ct.csv', package = 'pcr')
pcr1_ct <- readr::read_csv(fl)

# add grouping variable
group_var <- rep(c('brain', 'kidney'), each = 6)

# calculate averages
ave <- pcr_ave(pcr1_ct, group_var = group_var)

# calculate delta ct
pcr_dct(ave, 'GAPDH')

}
