% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_functions.R
\name{pcr_ddct}
\alias{pcr_ddct}
\title{Calculate delta delta ct value}
\usage{
pcr_ddct(df, reference_group)
}
\arguments{
\item{df}{A data.frame of delta ct values for each gene and a grouping
variable such as the output of the \link{pcr_dct}}

\item{reference_group}{A character string of the reference group as it is
recorded in the grouping variable}
}
\value{
A data.fram of the delta delta values for each gene in a grouping
variable
}
\description{
Takes a data.frame of the delta ct values for each gene in each condition
and returns a data.frame of the delta delta value of each gene (delat ct of
gene of interest - delta ct value of same gene in reference group)
}
\examples{
# locate and read raw ct data
fl <- system.file('extdata', 'pcr1_ct.csv', package = 'pcr')
pcr1_ct <- readr::read_csv(fl)

# add grouping variable
group_var <- rep(c('brain', 'kidney'), each = 6)

# calculate averages
ave <- pcr_ave(pcr1_ct, group_var = group_var)

# calculate delta ct
dct <- pcr_dct(ave, 'GAPDH')

# calculate delta delta ct
pcr_ddct(dct, 'brain')

}
