% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{pcr_dilute_error}
\alias{pcr_dilute_error}
\title{PCR dilution errors}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 2 columns.}
\usage{
pcr_dilute_error
}
\description{
PCR dilution errors
}
\references{
\url{http://www3.appliedbiosystems.com/cms/groups/mcb_support/documents/generaldocuments/cms_040980.pdf}
}
\keyword{datasets}
