% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_functions.R
\name{pcr_error}
\alias{pcr_error}
\title{Calculate error value}
\usage{
pcr_error(df, reference_gene)
}
\arguments{
\item{df}{A data.frame of ncol n and nrow equalse the number of unique
grouping variables containing the standard error values of each gene in each
group}

\item{reference_gene}{A character string of the name of the column
correspoinding to the reference gene}
}
\value{
A data.frame of error values for each gene and a grouping
variable. The column corresponding to the reference gene is dropped.
}
\description{
Calculate error value
}
