% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses_fun.R
\name{pcr_genorm}
\alias{pcr_genorm}
\title{Calculate the geNorm method}
\usage{
pcr_genorm(df, group_var, reference_group, reference_gene)
}
\arguments{
\item{df}{A data.frame of \eqn{C_T} values with genes in the columns and samples
in rows rows}

\item{group_var}{A character vector of a grouping variable. The length of
this variable should equal the number of rows of df}

\item{reference_group}{A character string of the control group in group_var}

\item{reference_gene}{A character string of the column name of a control gene}
}
\value{
A data.frame
}
\description{
Calculate the geNorm method
}
