% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{pcr_hk}
\alias{pcr_hk}
\title{Simulated data of two identical housekeeping genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 3 columns.}
\usage{
pcr_hk
}
\description{
Simulated data of two identical housekeeping genes
}
\keyword{datasets}
