% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{pcr_hk_calib}
\alias{pcr_hk_calib}
\title{Processed data of two identical houekeeping genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 7 columns.}
\usage{
pcr_hk_calib
}
\description{
Processed data of two identical houekeeping genes
}
\keyword{datasets}
