% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{pcr_m}
\alias{pcr_m}
\title{Calculate m values for reference genes}
\usage{
pcr_m(df, group_var, reference_gene)
}
\arguments{
\item{df}{A \code{data.frame} of the expression values.}

\item{group_var}{A \code{character} vector of the groups of samples of
length equals the number of rows in \code{df}.}

\item{reference_gene}{A \code{character} vector of the names of reference
genes in \code{df} column names.}
}
\value{
A \code{data.frame} of a number of colums equals the length of
\code{reference_gene} plus a \code{group} column and a number of rows equals
the number of unique values in \code{group_var}.
}
\description{
Calculates the stability measure for a number of reference genes
}
\examples{
## locate and read relative expression of reference genes
fl <- system.file('extdata', 'hk_tissue.csv', package = 'pcr')
hk_tissue <- readr::read_csv(fl)

# make a grouping variable and names of reference genes
group_var <- ifelse(grepl('POOL', hk_tissue$tissue),
    'Normal',
     hk_tissue$tissue)
reference_gene <- names(hk_tissue)[-11]

# calculate m values for all reference genes in all groups
pcr_m(hk_tissue, group_var, reference_gene)

}
