% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses_fun.R
\name{pcr_nf}
\alias{pcr_nf}
\title{Calculate normalization factors}
\usage{
pcr_nf(df, group_var, reference_group, reference_gene)
}
\arguments{
\item{df}{A data.frame of \eqn{C_T} values with genes in the columns and samples
in rows rows}

\item{group_var}{A character vector of a grouping variable. The length of
this variable should equal the number of rows of df}

\item{reference_group}{A character string of the control group in group_var}

\item{reference_gene}{A character string of the column name of a control gene}
}
\value{
A data.frame of 5 columns
\itemize{
  \item group The unique entries in group_var
  \item gene The column names of df. reference_gene is dropped
  \item mean The average value of the normalization factor
  \item sd The standard deviation of the normalization factor
  \item factor The name of the normalization factor 'NF'
}
}
\description{
Calculates a normalization factor from multiple reference genes.
}
\examples{
## locate and read raw ct data
fl <- system.file('extdata', 'ct5.csv', package = 'pcr')
ct5 <- readr::read_csv(fl)

# add grouping variable
group_var <- rep(c('control', 'treatment'), each = 2)

# calculate a normalization factor
pcr_nf(ct5,
       group_var = group_var,
       reference_group = 'control',
       reference_gene = paste0('REF', 1:3))

}
