% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fun.R
\name{pcr_plot_analyze}
\alias{pcr_plot_analyze}
\title{Plotting function}
\usage{
pcr_plot_analyze(df, method, facets = FALSE)
}
\arguments{
\item{df}{A data.frame}

\item{method}{A character string}

\item{facets}{A character string}
}
\value{
A plot
}
\description{
Plotting function
}
