% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fun.R
\name{pcr_plot_assess}
\alias{pcr_plot_assess}
\title{Plot quality assessment graphs}
\usage{
pcr_plot_assess(df, amount, reference_gene, method)
}
\arguments{
\item{df}{A data.frame}

\item{amount}{A numeric vector}

\item{reference_gene}{A character string}

\item{method}{A character string; 'efficiency' or 'standard_curve'}
}
\value{
A plot
}
\description{
Plot quality assessment graphs
}
