% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{pcr_stability}
\alias{pcr_stability}
\title{Calculate stability of a combination of reference genes}
\usage{
pcr_stability(df, group_var, reference_gene)
}
\arguments{
\item{df}{A \code{data.frame} of the expression values.}

\item{group_var}{A \code{character} vector of the groups of samples of
length equals the number of rows in \code{df}.}

\item{reference_gene}{A \code{character} vector of the names of reference
genes in \code{df} column names.}
}
\value{
A \code{data.frame} of a number of columns equals the number of
uniqure entries in \code{group_var} plus two other columns; \code{gene}
and \code{n} with the names and numbers of the combinations of
reference genes used.
}
\description{
Calculate the stability of a decreasing number of combinations of
reference genes by removing the least stable reference gene across all
groups.
}
\examples{
## locate and read relative expression of reference genes
fl <- system.file('extdata', 'hk_tissue.csv', package = 'pcr')
hk_tissue <- readr::read_csv(fl)

# make a grouping variable and names of reference genes
group_var <- ifelse(grepl('POOL', hk_tissue$tissue),
    'Normal',
     hk_tissue$tissue)
reference_gene <- names(hk_tissue)[-11]

# calculate m values for all reference genes in all groups
pcr_stability(hk_tissue, group_var, reference_gene)

}
