% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{pcr_trend}
\alias{pcr_trend}
\title{Calculate the linear trend for the standard curve}
\usage{
pcr_trend(df, amount)
}
\arguments{
\item{df}{A data.frame of numeric ct value of n genes in columns and m
samples in raws}

\item{amount}{A numeric vector of the input dilutions or amounts}
}
\value{
A data.frame of 4 columns
\itemize{
  \item gene
  \item intercept
  \item slope
  \item r_squared
}
}
\description{
Calculate the linear trend for the standard curve
}
\examples{
# locate and read file
fl <- system.file('extdata', 'ct3.csv', package = 'pcr')
ct3 <- readr::read_csv(fl)

# make amount/dilution variable
amount <- rep(c(1, .5, .2, .1, .05, .02, .01), each = 3)

# calculate trend
pcr_trend(ct3, amount = amount)

}
