% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{rel_express2}
\alias{rel_express2}
\title{PCR same wells experiment processed data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 8 columns.}
\usage{
rel_express2
}
\description{
PCR same wells experiment processed data
}
\keyword{datasets}
