% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{standarized2}
\alias{standarized2}
\title{PCR amount using standard curve/ average normalized mode}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 7 columns.}
\usage{
standarized2
}
\description{
PCR amount using standard curve/ average normalized mode
}
\keyword{datasets}
