% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{.pcr_amount}
\alias{.pcr_amount}
\title{Calculate PCR RNA amounts}
\usage{
.pcr_amount(df, intercept, slope)
}
\arguments{
\item{df}{A data.frame}

\item{intercept}{A numeric vector of length equals the number of columns of df,
one for each gene such as the output of \link{pcr_assess}}

\item{slope}{A numeric vector of length equals the number of columns of df,
one for each gene such as the output of \link{pcr_assess}}
}
\value{
A data.frame with the same dimensions as df containing the
amount of RNA in each sample in each gene
}
\description{
Calculate PCR RNA amounts
}
\details{
Used to alculate the amount of RNA in a PCR experimental sample using the
information provided by the standard curve, namely the slope and the
intercept calculated in advance for each gene in a similar expermiment.
}
\examples{
# locate and read data
fl <- system.file('extdata', 'ct3.csv', package = 'pcr')
ct3 <- read.csv(fl)

fl <- system.file('extdata', 'ct1.csv', package = 'pcr')
ct1 <- read.csv(fl)

# make a vector of RNA amounts
amount <- rep(c(1, .5, .2, .1, .05, .02, .01), each = 3)

# calculate curve
standard_curve <- pcr_assess(ct3, amount = amount, method = 'standard_curve')
intercept <- standard_curve$intercept
slope <- standard_curve$slope

# calculate amounts
.pcr_amount(ct1,
           intercept = intercept,
           slope = slope)

}
\keyword{internal}
