% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{.pcr_cv}
\alias{.pcr_cv}
\title{Calculates the coefficient of variation}
\usage{
.pcr_cv(amounts, group_var, tidy = FALSE)
}
\arguments{
\item{amounts}{A data.frame of the calculated input amounts returned by
\code{\link{.pcr_amount}}}

\item{tidy}{A logical, default FALSE. When TRUE returns a tidy data.frame}
}
\value{
A data.frame with a column for the grouping variable; group_var and
one for each of the original columns. When tidy is TRUE the
returns a tidy data.frame with 3 columns; group/amount, gene and error
}
\description{
Calculates the coefficient of variation of a gene by deviding standard
deviations of each group by their averages
}
\details{
Used to calculate the coefficient of variation of the input amounts
 after grouping by a grouping variable; group_var for experimental groups.
}
\examples{
# locate and read data
fl <- system.file('extdata', 'ct3.csv', package = 'pcr')
ct3 <- read.csv(fl)

fl <- system.file('extdata', 'ct1.csv', package = 'pcr')
ct1 <- read.csv(fl)

# make a vector of RNA amounts
amount <- rep(c(1, .5, .2, .1, .05, .02, .01), each = 3)

# calculate curve
standard_curve <- pcr_assess(ct3, amount = amount, method = 'standard_curve')
intercept <- standard_curve$intercept
slope <- standard_curve$slope

# calculate amounts
input_amounts <- .pcr_amount(ct1,
                            intercept = intercept,
                            slope = slope)

# make grouping variable
group <- rep(c('brain', 'kidney'), each = 6)

# calculate cv errors
.pcr_cv(input_amounts,
       group_var = group)

}
\keyword{internal}
