% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{.pcr_normalize}
\alias{.pcr_normalize}
\title{Normalize vector by another}
\usage{
.pcr_normalize(vec, ref, mode = "subtract")
}
\arguments{
\item{vec}{A vector of numerics}

\item{ref}{A numeric vector}

\item{mode}{Either 'subtract' or 'divide'}
}
\value{
A vector of numerics
}
\description{
Normalize vector by another
}
\examples{
vec <- rnorm(6, 30, 1)
ref <- rnorm(6, 30, .1)
pcr:::.pcr_normalize(vec, ref)
pcr:::.pcr_normalize(vec, ref, mode = 'divide')

}
\keyword{internal}
