% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{.pcr_sd}
\alias{.pcr_sd}
\title{Calculate standard deviation}
\usage{
.pcr_sd(df, group_var, tidy = FALSE)
}
\arguments{
\item{df}{A data.frame}

\item{group_var}{A vector or lenght equals the number or rows of df}

\item{tidy}{A logical, default FALSE. When TRUE returns a tidy data.frame}
}
\value{
A data.frame with a column for the grouping variable; group_var and
one for each of the original columns. When tidy is TRUE the
returns a tidy data.frame with 3 columns; group/amount, gene and error.
}
\description{
Uses a group_by statement to calculate the standard deviations of values in
a data.fram grouped by a variable
}
\details{
Used to calculate the standard devaitions of ct after grouping by a
group_var for the experimental groups
}
\examples{
# locate and read raw ct data
fl <- system.file('extdata', 'ct1.csv', package = 'pcr')
ct1 <- readr::read_csv(fl)

# add grouping variable
group_var <- rep(c('brain', 'kidney'), each = 6)

# calculate standard deviations
.pcr_sd(ct1, group_var = group_var)

# calculate standard deviations and return a tidy data.frame
.pcr_sd(ct1, group_var = group_var, tidy = TRUE)

}
\keyword{internal}
