% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\docType{package}
\name{pcr}
\alias{pcr}
\title{\code{pcr} package}
\description{
Analyzing real-time quantitative PCR data
}
\details{
Calculates the amplification efficiency and curves from real-time
quantitative PCR (Polymerase Chain Reaction) data. Estimates the relative
expression from PCR data using the double delta CT and the standard curve
methods Livak & Schmittgen (2001) <doi:10.1006/meth.2001.1262>. Tests for
statistical significance using two-group tests and linear regression
Yuan et al. (2006) <doi: 10.1186/1471-2105-7-85>.
}
