% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_functions.R
\name{pcr_amount}
\alias{pcr_amount}
\title{Calculate PCR RNA amounts}
\usage{
pcr_amount(df, intercept, slope)
}
\arguments{
\item{df}{A data.frame of numeric ct value of n genes in columns and m
samples in raws.}

\item{intercept}{A numeric vector of length equals ncol(df), one for each
gene such as the output of \link{pcr_assess}}

\item{slope}{A numeric vector of length ncol(df), one for each gene such as
the output of \link{pcr_assess}}
}
\value{
A data.frame of dimensions equal that of the input df containing the
amount of RNA in each sample in each gene
}
\description{
Calculate the amount of RNA in a PCR experimental sample using the
information provided by the standard curve, namely the slope and the
intercept calculated in advance for each gene in a similar expermiment.
}
\examples{
# locate and read data
fl <- system.file('extdata', 'pcr_dilute.csv', package = 'pcr')
pcr_dilute <- readr::read_csv(fl)

fl <- system.file('extdata', 'pcr1_ct.csv', package = 'pcr')
pcr1_ct <- readr::read_csv(fl)

# make a vector of RNA amounts
amount <- rep(c(1, .5, .2, .1, .05, .02, .01), each = 3)

# calculate curve
standard_curve <- pcr_assess(pcr_dilute, amount = amount, mode = 'standard_curve')
intercept <- standard_curve$intercept
slope <- standard_curve$slope

# calculate amounts
pcr_amount(pcr1_ct,
           intercept = intercept,
           slope = slope)

}
