% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses_fun.R
\name{pcr_analyze}
\alias{pcr_analyze}
\title{Apply qPCR analysis methods}
\usage{
pcr_analyze(df, method = "delat_delat_ct", ...)
}
\arguments{
\item{df}{A data.frame of numeric ct value of n genes in columns and m
samples in raws}

\item{method}{A character string; 'delta_delta_ct' defalult, 'delta_ct' or
'relative_curve'}

\item{...}{Arguments passed to other methods}
}
\value{
A data.frame or a plot. For details; \link{pcr_ddct}, \link{pcr_dct} and
\link{pcr_curve}
}
\description{
Apply qPCR analysis methods
}
