% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_functions.R
\name{pcr_assess}
\alias{pcr_assess}
\title{Assess PCR data quality}
\usage{
pcr_assess(df, amount, mode = "effeciency", plot = FALSE)
}
\arguments{
\item{df}{A data.frame of one or more columns containing ct values of a
reference gene and another from an experiment run with differen dilutions
(RNA amounts)}

\item{amount}{A numeric vector of length equals nrow df with RNA amounts}

\item{mode}{A character string of assessment mode. Default "effeciency"}

\item{plot}{A logical default FALSE of whether plot or return the data}
}
\value{
When mode == 'effeciency' return a data.frame of 5 columns or a plot
. when mode == 'standard_curve' retruns a data.frame of intercept and slope
or a plot for each of the input columns.
}
\description{
Assess the quality of a PCR experiment by examining the amplification
effeciency when using the comparative ct method or calculating the standard
curve
}
\examples{
# locate and read data
fl <- system.file('extdata', 'pcr_dilute.csv', package = 'pcr')
pcr_dilute <- readr::read_csv(fl)

# make a vector of RNA amounts
amount <- rep(c(1, .5, .2, .1, .05, .02, .01), each = 3)

# calculate effeciencey
pcr_assess(pcr_dilute,
           amount = amount)

# calculate standard curve
pcr_assess(pcr_dilute,
           amount = amount,
           mode = 'standard_curve')

}
