% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_fun.R
\name{pcr_assess}
\alias{pcr_assess}
\title{Assess qPCR data quality}
\usage{
pcr_assess(df, method = "standard_curve", ...)
}
\arguments{
\item{df}{A data.frame of numeric ct value of n genes in columns and m
samples in raws}

\item{method}{A character string; 'standard_curve' defalult or 'effeciency'}

\item{...}{Arguments passed to other methods}
}
\value{
A data.frame or a plot. For details; \link{pcr_standard} and
\link{pcr_effeciency}
}
\description{
Assess qPCR data quality
}
