% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{pcr_caliberate}
\alias{pcr_caliberate}
\title{Caliberate ct values to a reference group}
\usage{
pcr_caliberate(df, reference_group, mode = "subtract", tidy = FALSE)
}
\arguments{
\item{df}{A data.frame of ct or normalized ct values for each gene and a
grouping variable such as the output of the \link{pcr_normalize}}

\item{reference_group}{A character string of the reference group as it is
recorded in the grouping variable}

\item{mode}{A character string of the normalization mode to be used. Default
is 'subtract'. Other possible modes include 'divide'}

\item{tidy}{A logical, default FALSE. When TRUE returns a tidy data.frame}
}
\value{
A data.fram of the caliberated ct values for each gene in a grouping
variable by a reference group
}
\description{
Takes a data.frame of the ct or normalized ct values for each gene in each
condition and returns a data.frame of the caliberated ct value of each gene
e.x. (delat ct of gene of interest - delta ct value of same gene in
reference group)
}
\examples{
# locate and read raw ct data
fl <- system.file('extdata', 'ct1.csv', package = 'pcr')
ct1 <- readr::read_csv(fl)

# add grouping variable
group_var <- rep(c('brain', 'kidney'), each = 6)

# calculate averages
ave <- pcr_average(ct1, group_var = group_var)

# calculate delta ct
dct <- pcr_normalize(ave, 'GAPDH')

# calculate delta delta ct
pcr_caliberate(dct, 'brain', tidy = TRUE)

# calculate delta delta ct and return a tidy data.frame
pcr_caliberate(dct, 'brain', tidy = TRUE)

}
