% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{pcr_cv}
\alias{pcr_cv}
\title{Calculates cv}
\usage{
pcr_cv(amounts, group_var, reference_gene, mode = "average_amount",
  tidy = FALSE)
}
\arguments{
\item{amounts}{A data.frame of calculated input RNA amounts}

\item{group_var}{A vector of length m as a grouping variables of samples}

\item{reference_gene}{A character string of the name of the column
corresponding to the reference gene}

\item{mode}{A character string; 'separate_tube' or 'same_tube'}

\item{tidy}{A logical, default FALSE. When TRUE returns a tidy data.frame}
}
\value{
A data.frame of the cv term for each gene in each sample
}
\description{
Error terms for the standard curve relative quantification
}
\examples{
# locate and read data
fl <- system.file('extdata', 'ct3.csv', package = 'pcr')
ct3 <- readr::read_csv(fl)

fl <- system.file('extdata', 'ct1.csv', package = 'pcr')
ct1 <- readr::read_csv(fl)

# make a vector of RNA amounts
amount <- rep(c(1, .5, .2, .1, .05, .02, .01), each = 3)

# calculate curve
standard_curve <- pcr_assess(ct3, amount = amount, method = 'standard_curve')
intercept <- standard_curve$intercept
slope <- standard_curve$slope

# calculate amounts
input_amounts <- pcr_amount(ct1,
                            intercept = intercept,
                            slope = slope)

# make grouping variable
group <- rep(c('brain', 'kidney'), each = 6)

# calculate cv errors
pcr_cv(input_amounts,
       group_var = group,
       reference_gene = 'GAPDH')

}
