% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses_fun.R
\name{pcr_dct}
\alias{pcr_dct}
\title{Calculate the delta_ct}
\usage{
pcr_dct(df, group_var, reference_gene, reference_group, mode = "average_ct")
}
\arguments{
\item{df}{A data.frame of ct values}

\item{group_var}{A character vector of a grouping variable}

\item{reference_gene}{A character string of the column name of a control gene}

\item{reference_group}{A character string of the control group}

\item{mode}{A character string of; 'average_ct' or 'average_dct'}
}
\value{
A data.frame of 7 columns
\itemize{
  \item group
  \item gene
  \item caliberated
  \item fold_change
  \item error
  \item lower
  \item upper
}
}
\description{
Calculate the delta_ct
}
\examples{
# locate and read file
fl <- system.file('extdata', 'ct1.csv', package = 'pcr')
ct1 <- readr::read_csv(fl)

# make a data.frame of two identical columns
pcr_hk <- data.frame(
  GAPDH1 = ct1$GAPDH,
  GAPDH2 = ct1$GAPDH
  )

group_var <- rep(c('brain', 'kidney'), each = 6)

# calculate caliberation
pcr_dct(pcr_hk,
        group_var = group_var,
        reference_group = 'brain')

}
