% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses_fun.R
\name{pcr_ddct}
\alias{pcr_ddct}
\title{Calculate the delta_delta_ct}
\usage{
pcr_ddct(df, group_var, reference_gene, reference_group, mode = "average_ct")
}
\arguments{
\item{df}{A data.frame of ct values}

\item{group_var}{A character vector of a grouping variable}

\item{reference_gene}{A character string of the column name of a control gene}

\item{reference_group}{A character string of the control group}

\item{mode}{A character string of; 'average_ct' or 'average_dct'}
}
\value{
A data.frame of 8 columns
\itemize{
  \item group
  \item gene
  \item normalized
  \item caliberated
  \item relative_expression
  \item error
  \item lower
  \item upper
}
}
\description{
Calculate the delta_delta_ct
}
\examples{
## locate and read raw ct data
fl <- system.file('extdata', 'ct1.csv', package = 'pcr')
ct1 <- readr::read_csv(fl)

# add grouping variable
group_var <- rep(c('brain', 'kidney'), each = 6)

# calculate all values and errors in one step
pcr_ddct(ct1,
         group_var = group_var,
         reference_gene = 'GAPDH',
         reference_group = 'brain')

}
