% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{pcr_dilute}
\alias{pcr_dilute}
\title{Regenerated PCR ct values of different dilutions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 2 columns.}
\usage{
pcr_dilute
}
\description{
Regenerated PCR ct values of different dilutions
}
\references{
\url{http://www3.appliedbiosystems.com/cms/groups/mcb_support/documents/generaldocuments/cms_040980.pdf}
}
\keyword{datasets}
