% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{pcr_effeciency}
\alias{pcr_effeciency}
\title{PCR effecienty calculated from pcr_dilute}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 10 columns.}
\usage{
pcr_effeciency
}
\description{
PCR effecienty calculated from pcr_dilute
}
\keyword{datasets}
