% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_fun.R
\name{pcr_effeciency}
\alias{pcr_effeciency}
\title{Calculate amplification effeciency}
\usage{
pcr_effeciency(df, amount, reference_gene, plot = FALSE)
}
\arguments{
\item{df}{A data.frame of numeric ct value of n genes in columns and m
samples in raws}

\item{amount}{A numeric vector of the input dilutions or amounts}

\item{reference_gene}{A character string of the name of the column
correspoinding to the reference gene}

\item{plot}{A logical (default FALSE)}
}
\value{
A data.frame of 4 columns when plot == FALSE. Returned by calling
\link{pcr_trend}
}
\description{
Calculate amplification effeciency
}
\examples{
# locate and read file
fl <- system.file('extdata', 'ct3.csv', package = 'pcr')
ct3 <- readr::read_csv(fl)

# make amount/dilution variable
amount <- rep(c(1, .5, .2, .1, .05, .02, .01), each = 3)

# calculate amplification effeciency
pcr_effeciency(ct3,
               amount = amount,
               reference_gene = 'GAPDH')

# plot amplification effeciency
pcr_effeciency(ct3,
               amount = amount,
               reference_gene = 'GAPDH',
               plot = TRUE)

}
