% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{pcr_error}
\alias{pcr_error}
\title{Calculate error value}
\usage{
pcr_error(df, reference_gene, tidy = FALSE)
}
\arguments{
\item{df}{A data.frame of ncol n and nrow equalse the number of unique
grouping variables containing the standard error values of each gene in each
group}

\item{reference_gene}{A character string of the name of the column
correspoinding to the reference gene}

\item{tidy}{A logical, default FALSE. When TRUE returns a tidy data.frame}
}
\value{
A data.frame of error values for each gene and a grouping
variable. The column corresponding to the reference gene is dropped.
}
\description{
Calculate error value
}
\examples{
# locate and read raw ct data
fl <- system.file('extdata', 'ct1.csv', package = 'pcr')
ct1 <- readr::read_csv(fl)

# add grouping variable
group_var <- rep(c('brain', 'kidney'), each = 6)

# calculate standard deviations
sds <- pcr_sd(ct1, group_var = group_var)

# calculate errors
pcr_error(sds, reference_gene = 'GAPDH')

}
