% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{pcr_normalize}
\alias{pcr_normalize}
\title{Normalize ct values to a reference gene}
\usage{
pcr_normalize(df, reference_gene, mode = "subtract", tidy = FALSE)
}
\arguments{
\item{df}{A data.frame of such as that returned by \link{pcr_average}}

\item{reference_gene}{A character string of the name of the column
correspoinding to the reference gene}

\item{mode}{A character string of the normalization mode to be used. Default
is 'subtract'. Other possible modes include 'divide'}

\item{tidy}{A logical, default FALSE. When TRUE returns a tidy data.frame}
}
\value{
A data.frame of normalized ct values for each gene and a
grouping variable. When mode is 'subtract' (default) each the reference gene
 is subtracted from each column, and division is used instead when mode is
'divide'. The functino returns ignors non numeric columns and dropes the one
that corresponds to the reference_gene
}
\description{
Takes a data.frame of ct values or average ct values for each gene in each
condition and returns the ct values normalized to a reference gene.
}
\examples{
# locate and read raw ct data
fl <- system.file('extdata', 'ct1.csv', package = 'pcr')
ct1 <- readr::read_csv(fl)

# normalize the ct values
pcr_normalize(ct1, reference_gene = 'GAPDH')

# normalize by division
pcr_normalize(ct1, reference_gene = 'GAPDH', mode = 'divide')

# add grouping variable and average first
group_var <- rep(c('brain', 'kidney'), each = 6)
ave <- pcr_average(ct1, group_var = group_var)

# normalize by subtraction
pcr_normalize(ave, 'GAPDH')

# normalize by division
pcr_normalize(ave, 'GAPDH', mode = 'divide')

}
