% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_functions.R
\name{pcr_normalize}
\alias{pcr_normalize}
\title{Normalize PCR experment data}
\usage{
pcr_normalize(df, group_var, reference_gene, reference_group,
  intervals = TRUE)
}
\arguments{
\item{df}{A data.frame of numeric ct value of n genes in columns and m
samples in raws.}

\item{group_var}{A vector of length m as a grouping variables of samples}

\item{reference_gene}{A character string of the name of the column
correspoinding to the reference gene}

\item{reference_group}{A character string of the reference group as it is
recorded in the grouping variable}

\item{intervals}{A logical (default TRUE) to whether or not to calculate the
error intervals}
}
\value{
A tidy data.frame of calculated normalized relative values and
errors
}
\description{
A wrapper function to calculate the normalized relative values of genes of
interest to a reference gene and a reference group.
}
\examples{
# locate and read raw ct data
fl <- system.file('extdata', 'pcr1_ct.csv', package = 'pcr')
pcr1_ct <- readr::read_csv(fl)

# add grouping variable
group_var <- rep(c('brain', 'kidney'), each = 6)

# calculate all values and errors in one step
pcr_normalize(pcr1_ct,
group_var = rep(c('brain', 'kidney'), each = 6),
reference_gene = 'GAPDH',
reference_group = 'brain')

}
