% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_functions.R
\name{pcr_sd}
\alias{pcr_sd}
\title{Calculate standard error value}
\usage{
pcr_sd(df, group_var)
}
\arguments{
\item{df}{A data.frame of numeric ct value of n genes in columns and m
samples in raws.}

\item{group_var}{A vector of length m as a grouping variables of samples}
}
\value{
A data.frame of ncol n and nrow equalse the number of unique
grouping variables containing the standard error values of each gene in each
group
}
\description{
Calculate standard error value
}
\examples{
# locate and read raw ct data
fl <- system.file('extdata', 'pcr1_ct.csv', package = 'pcr')
pcr1_ct <- readr::read_csv(fl)

# add grouping variable
group_var <- rep(c('brain', 'kidney'), each = 6)

# calculate standard deviations
pcr_sd(pcr1_ct, group_var = group_var)

}
