% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_fun.R
\name{pcr_test}
\alias{pcr_test}
\title{Statistical testing of PCR data}
\usage{
pcr_test(df, group_var, reference_gene, test = "t.test")
}
\arguments{
\item{df}{A data.frame of numeric ct value of n genes in columns and m
samples in raws}

\item{group_var}{A vector of length m as a grouping variables of samples}

\item{test}{A character string; 't.test' default or 'wilcox.test'}
}
\value{
A data.frame of 5 columns
\itemize{
  \item gene
  \item estimate
  \item p_value
  \item lower
  \item upper
}
}
\description{
Statistical testing of PCR data
}
\examples{
# locate and read data
fl <- system.file('extdata', 'ct4_raw.tsv', package = 'pcr')
ct4 <- read_csv(fl)

# make group variable
group <- rep(c('control', 'treatment'), each = 12)
group <- relevel(factor(group), ref = 'treatment')

# test using t-test
pcr_test(ct4,
         group_var = group,
         reference_gene = 'ref',
         test = 't.test')

# test using wilcox.test
pcr_test(ct4,
         group_var = group,
         reference_gene = 'ref',
         test = 'wilcox.test')

}
